adds REST call to persona for subscription
fixes the history not being passed in
